third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/video-intrinsic-width-height-expected.txt

This is a testharness.js-based test.
[FAIL] only width attribute
  assert_equals: height expected 50 but got 150
[FAIL] only height attribute
  assert_equals: width expected 200 but got 300
[FAIL] both width/height attributes and style keeping aspect-ratio
  assert_equals: width expected 300 but got 500
Harness: the test ran to completion.
